+++ Sources/KeyboardKit/_Essentials/Keyboard+StandardBehavior.swift
@@ -29,20 +29,21 @@ extension Keyboard {
         /// - Parameters:
         ///   - keyboardContext: The keyboard context to use.
         ///   - doubleTapThreshold: The douple tap threshold to use, by default `0.5` seconds.
+        ///   - endSentenceText: The text to use to end sentences with, by default `. `.
         ///   - endSentenceThreshold: The end sentence auto-close threshold to use, by default `3.0` seconds.
         ///   - repeatGestureTimer: The repease gesture timer to use, by default ``Gestures/RepeatTimer/shared``.
         public init(
             keyboardContext: KeyboardContext,
             doubleTapThreshold: TimeInterval = 0.5,
+            endSentenceText: String = ". ",
             endSentenceThreshold: TimeInterval = 3.0,
-            repeatGestureTimer: Gestures.RepeatTimer = .shared,
-            endSentenceSymbolic: String = ". "
+            repeatGestureTimer: Gestures.RepeatTimer = .shared
         ) {
             self.keyboardContext = keyboardContext
             self.doubleTapThreshold = doubleTapThreshold
+            self.endSentenceText = endSentenceText
             self.endSentenceThreshold = endSentenceThreshold
             self.repeatGestureTimer = repeatGestureTimer
-            self.endSentenceSymbolic = endSentenceSymbolic
         }
         
         
@@ -51,7 +52,10 @@ extension Keyboard {
         
         /// The douple tap threshold to use.
         public let doubleTapThreshold: TimeInterval
-        
+
+        /// The text to use to end sentences with.
+        public var endSentenceText: String
+
         /// The end sentence auto-close threshold to use.
         public let endSentenceThreshold: TimeInterval
         
@@ -73,9 +77,6 @@ extension Keyboard {
             let duration = repeatGestureTimer.duration ?? 0
             return duration > 3 ? .word : .character
         }
-        
-        /// The  symbolic that  end the sentence after a gesture action should  inject.
-        public var endSentenceSymbolic: String
         
         /// The preferred keyboard type after an action gesture.
         open func preferredKeyboardType(